automation/pyproject.toml
[project]
name = "automation"
version = "0.1.0"
description = "automation using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.108.0,<1.0.0", # Reverted as requested
    "pypdf>=5.0.0,<6.0.0",          # Updated requirement based on crewai dependency
    "pdfplumber>=0.11.4,<0.12.0",    # Compatible version for crewai>=0.108.0
    "python-dotenv>=1.0.0,<2.0.0",  # For loading .env files
    "google-generativeai>=0.4.0,<0.5.0" # Re-adding Gemini LLM dependency
]

[project.scripts]
automation = "automation.main:run"
run_crew = "automation.main:run"
train = "automation.main:train"
replay = "automation.main:replay"
test = "automation.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
